std: refactor svs spec package
